+++ innopoints/views/__init__.py
@@ -2,5 +2,6 @@ from .authentication import *
 from .project import *
 from .activity import *
 from .file import *
+from .product import *
 
 from .views import *  # temporary, until all views have been moved
